Report for end of year project - Word Document
This is a detailed report for the end of year project for HNC database technologies (2016). It details the design and building of a database for a Church organisation, following the SDLC using the waterfall methodology. It includes entity relationship modelling, intuitively breaking down of the tables to reduce redundancy, precedence tables, data dictionaries, scheduling of the entire project and any subsequent acceptance testing.